edge-modules/temp-simulator/requirements.txt
azure-iot-device
pandas